requirements.txt
Flask==3.1.1
spotipy==2.23.0
python-dotenv==1.0.1
